+++ mostlyai/__init__.py
@@ -1,4 +1,4 @@
-from mostlyai.api import MostlyAI
+from mostlyai.client.api import MostlyAI
 
 __all__ = ["MostlyAI"]
 __version__ = "0.4.10"  # Do not set this manually. Use poetry version [params].
